backend/e2ee_chat/urls.py
from django.contrib import admin
from django.urls import path, include
from rest_framework import routers

from chat.views import ConversationViewSet, MessageViewSet
from users.views import UserViewSet

router = routers.DefaultRouter()
router.register(r'users', UserViewSet)
router.register(r'conversations', ConversationViewSet)
router.register(r'messages', MessageViewSet)

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/', include(router.urls)),
    path('api/auth/', include('rest_framework.urls')),
]
