backend/requirements.txt
Django>=4.2
djangorestframework
channels
channels-redis
daphne
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
